create initContainer for loading snapshots
if key is set, download and untar a blocks directory into the bitcoin
datadir. its up to the user to make sure they are choosing an
appropriate snapshot, e.g., if the snapshot was created from a signet
chain then it can only be loaded into a signet node with the correct
signet challenge. we can expand on this in the future to better export
all of the relevant config details in the snapshot, but for now this
just adds the ability to load files into pods before the pod starts.